requirements.txt
# Core Django
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-environ==0.11.2

# Database
psycopg2-binary==2.9.9

# Authentication
djangorestframework-simplejwt==5.3.0
PyJWT==2.8.0

# API Documentation
drf-spectacular==0.27.1

# Background Tasks
celery==5.3.4
redis==5.0.1

# Error Tracking (optional)
# sentry-sdk==1.40.0  # Uncomment and set SENTRY_DSN if using Sentry

# HTTP Client for Webhooks
requests==2.31.0

# Testing
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
faker==20.1.0

# Utilities
python-dotenv==1.0.0

